% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_naming.R
\name{.capitalize_medical_term}
\alias{.capitalize_medical_term}
\title{Capitalize medical terms appropriately}
\usage{
.capitalize_medical_term(term)
}
\description{
Capitalize medical terms appropriately
}
\keyword{internal}
