% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format.R
\name{val_p_format}
\alias{val_p_format}
\title{Format a p-value for reporting}
\usage{
val_p_format(p, digits = 3)
}
\arguments{
\item{p}{Numeric p-value}

\item{digits}{Number of decimal places to round to (default 3)}
}
\value{
Formatted string with specified decimals, scientific if < 0.001
}
\description{
Format a p-value for reporting
}
