% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format.R
\name{val_format}
\alias{val_format}
\title{Format a mean  +-  SD string}
\usage{
val_format(mean, sd)
}
\arguments{
\item{mean}{Mean value}

\item{sd}{Standard deviation}
}
\value{
Formatted string "mean  +-  sd"
}
\description{
Format a mean  +-  SD string
}
