% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_methods_doc.r
\name{write_methods_doc}
\alias{write_methods_doc}
\title{Write methods section document for ternG output}
\usage{
write_methods_doc(tbl, filename, n_levels = 2, OR_col = FALSE)
}
\arguments{
\item{tbl}{A tibble created by ternG}

\item{filename}{Output file path ending in .docx}

\item{n_levels}{Number of group levels (2 for two-group, 3+ for multi-group)}

\item{OR_col}{Logical; whether odds ratios were calculated}
}
\description{
Write methods section document for ternG output
}
